This is sort of "Version 0" of the Organism domain class. This is basically all of the metadata we get from processing the raw genbank files. If we end up processing more data, we can add more later, but this amount of data fulfills the base requirements for the project.
Notice the Map type for the taxonomy field, and it is not required and in fact, nullable. Not all organisms in genbank have taxonomy info, so when we add the data we will just add a key --> value mapping from rank (i.e. kingdom) to value (i.e. eukaryota) and store the values that way.

setting "static scaffold = true" in the OrganismController tells Grails to make the basic GUI for Create, Read, Update, Delete.

The OrganismTest unit test verifies that the constraints I have set behave the way I expect them to.